requirements.txt
fastapi
uvicorn
google-cloud-storage
python-multipart
pydantic-settings